SideBar Refactoring Updates 1. feat: create SidebarHeader, SidebarNav, and SidebarFooter components 2. feat: create AppTitle component for sidebar branding 3. feat: create ThemeSelector component for theme switching 4. feat: create CollapseButton component for sidebar toggle 5. feat: create LogoutButton component for user logout 6. refactor: update SidebarHeader to use abstracted components 7. refactor: update SidebarFooter to use LogoutButton component
