+++ penny-quest-frontend/src/app/components/sideBarComponents/SidebarFooter.tsx
@@ -1,0 +1,17 @@
+import LogoutButton from './LogoutButton';
+
+interface SidebarFooterProps {
+  isCollapsed: boolean;
+}
+
+export default function SidebarFooter({ isCollapsed }: SidebarFooterProps) {
+  return (
+    <>
+      {!isCollapsed && (
+        <div className="px-6 pb-6">
+          <LogoutButton />
+        </div>
+      )}
+    </>
+  );
+}
